requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.4.0
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.17.0
python-dotenv==1.0.1
aiofiles==23.2.1
